Задание 6. Доделала задание со статичным классом

+++ src/Director.java
@@ -2,12 +2,12 @@ public class Director {
     int workers;
     Worker worker = new Worker();
 
-    public double getNumberOfSubordinates() {
+    public int getNumberOfSubordinates() {
         return workers;
     }
 
     public double getSalary() {
-        double stavka = worker.getSalary() * (getNumberOfSubordinates() / 100 * 9);
+        double stavka = worker.getSalary() * ((double) getNumberOfSubordinates() / 100 * 9);
         return getNumberOfSubordinates() == 0 ? worker.getSalary() : stavka;
     }
 }
